backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
pydantic==2.10.3
python-multipart==0.0.12
python-dotenv==1.0.1
pypdf==4.3.1
python-docx==1.1.2
